requirements.txt
click
requests
pysensu-ng
